Add files via upload
I added 3 rooms 2 of them are connected to the Starting room (purple and blue) the "Last Room" is locked by not having any doors and the only way to get in is the "Key"

An ending I added in relation to my rooms is if you enter the "Last Room" by having the Key in your inventory you get a message and win the game. I also added ending if you try to add any of my 3 dragons to your inventory.

My locked door is the "Last Room" it is locked because it does not have any doors to connect it but if you have the key in your inventory which you have to pick up then you go into the room.

I added 3 subclasses such as Dragon, Key, Phone, I had to override the toString method so things properly appeared in my inventory, dragon also has endings with them, and you can do a 6th action if you have the phone which is calling your mother who is another person I added to the game.

I added 3 different dragons and have different endings for each, 1 you fly away and live a happy life. 2 you learn you were its master and reconnect with it. 3 it eats you deals damage to you and you die.

I added 2 tests 1 of each heals you and the other makes you take damage. 

I added health to the person and he takes damage when the dragon eats him. But there is no way to fight the dragon.